requirements.txt
streamlit
opencv-python
face_recognition
Pillow
numpy
